ImageBind/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
# torch==1.13.0
torchvision==0.14.0
ftfy
regex
einops
fvcore
decord==0.6.0
iopath
numpy
matplotlib
types-regex
mayavi
lightning==2.0.2
cartopy
safetensors
scikit-learn
timm==0.6.7
torchaudio==0.13.0
pytorchvideo @ git+https://github.com/facebookresearch/pytorchvideo.git@28fe037d212663c6a24f373b94cc5d478c8c1a1d

